% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_functions.R
\name{check_between}
\alias{check_between}
\title{Check if a time falls within a window}
\usage{
check_between(time, start, end)
}
\arguments{
\item{time}{the time to check against}

\item{start}{the start of the window}

\item{end}{the end of the window}
}
\value{
boolean, whether the time falls inside the window.
}
\description{
Check if a time falls within a window
}

