% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{Pretty cross-tabulation}
\usage{
crosstable(df, var1, var2, weight_var = NULL, margins = FALSE,
  percent = FALSE)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{var1}{a field in \code{df} that will make the rows of the table.}

\item{var2}{a field in \code{df} that will make the columns of the table.}

\item{weight_var}{a field in \code{df} that weights the data. If
\code{NULL}, will treat each record equally.}

\item{margins}{boolean, should the row and column sums be appended?
default \code{FALSE}}

\item{percent}{boolean, should the table be a percentage?
default \code{FALSE}}
}
\value{
an array that can be cast with \code{kable()} or examined directly.
}
\description{
Pretty cross-tabulation
}

