% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{determine_linkout}
\alias{determine_linkout}
\title{Determine whether to link out the activity.}
\usage{
determine_linkout(purpose, start, end, duration)
}
\arguments{
\item{purpose}{Trip purpose vector}

\item{start}{Trip start time}

\item{end}{Trip end time}

\item{duration}{Length of time in minutes for minimum linkout time.}
}
\value{
boolean vector, TRUE if the record should be dropped.
}
\description{
Work trip models can be thwarted unless we remove incidental stops from them.
This function flags activities to drop based on the duration of an activity
and the activity purposes surrounding it.
}

