% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{get_activity_type}
\alias{get_activity_type}
\title{Get activity types}
\usage{
get_activity_type(codes)
}
\arguments{
\item{codes}{a vector of activity codes as found in the NHTS \code{WHYTO} or
  \code{WHYFROM} variables.}
}
\value{
a vector of activity codes using the following types:
  \code{home, work, school, pudo, other}
}
\description{
This function requires a `home_codes`
object be included in the project environment.
}

