% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_functions.R
\name{get_duration}
\alias{get_duration}
\title{Calculate duration of activity or trip in minutes}
\usage{
get_duration(start, end)
}
\arguments{
\item{start}{a vector of start times in military time.}

\item{end}{a vector of end times in military time.}
}
\value{
the elapsed minutes between the times
}
\description{
Calculate duration of activity or trip in minutes
}

