% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_functions.R
\name{get_military_time}
\alias{get_military_time}
\title{Turn number of minutes after 4 AM into military time.}
\usage{
get_military_time(x)
}
\arguments{
\item{x}{a numeric vector representing the number of minutes elapsed since
4 am.}
}
\value{
a character vector expressing the number of elapsed minutes in
military time.
}
\description{
This is precisely the inverse of \code{get_mins_from_4am}.
}

