% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hh_functions.R
\name{get_segment}
\alias{get_segment}
\title{Get the market segment}
\usage{
get_segment(vehicles, people, type)
}
\arguments{
\item{vehicles}{vector with the number of vehicles in the household.}

\item{people}{vector with the number of people to compare against the number
  of vehicles.}
}
\value{
a vector with the market segment the household belongs to.
}
\description{
Get the market segment
}

