% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{read_trips}
\alias{read_trips}
\title{Read a trips data frame}
\usage{
read_trips(file, clean = TRUE, linkout = NULL)
}
\arguments{
\item{file}{the path to the data file}

\item{clean}{If TRUE, will reduce to regularly used columns and process mode 
codes into named modes.}

\item{linkout}{If given, will link out activities of less than this duration 
that are part of a home-to-work tour portion.}
}
\value{
A data frame with
}
\description{
Read a trips data frame
}

