% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{swap_pa}
\alias{swap_pa}
\title{Swap P and A}
\usage{
swap_pa(df)
}
\arguments{
\item{df}{a trips record data_frame}
}
\value{
a data_frame with the p and a block and taz appended
}
\description{
This function determines if O and D need to be swapped to create a P and A
record, and swaps them if necessary.
This function requires a `home_codes`
object be included in the project environment.
}

