% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{get_trip_purpose}
\alias{get_trip_purpose}
\title{Get the trip purposes}
\usage{
get_trip_purpose(o_purp, d_purp)
}
\arguments{
\item{o_purp}{a vector of origin purposes coded as in the NHTS \code{WHYTO}
variable}

\item{d_purp}{a vector of destination purposes coded as in the NHTS
  \code{WHYFROM} variable}
}
\value{
a vector of trip purposes on the following classification:
  \code{hbw, hbo, hbschool, nhbo, nhbw}
}
\description{
This function requires `home_codes`, `work_codes`, and `school_codes`
objects be included in the project environment.
}

